refactor: move LotteryAgent
